refactor: 모임생성 타입 분리 ##64
