+++ src/app/(pages)/(club)/club/_api/supabase.ts
@@ -1,23 +1,18 @@
 import browserClient from "@/utils/supabase/client";
-import {
-  InsertWishList,
-  OneTimeClubForm,
-  OneTimeMember,
-  RegularClubForm,
-  RegularMember,
-  RegularRequest
-} from "../_types/ClubForm";
-import { InsertNotificationMember, RegularClubNotification } from "../regular-club-sub/[id]/create/_types/subCreate";
+
+import { EggDay, InsertEggDayMember } from "@/types/eggday.types";
+import { EggPopForm, EggPopMember } from "@/types/eggpop.types";
+import { EggClubForm, EggClubMember, EggClubRequest, InsertWishList } from "@/types/eggclub.types";
 
 // supabase에 일회성 모임 제출
-export const submitOneTimeClubData = async (finalFormData: OneTimeClubForm) => {
+export const submitOneTimeClubData = async (finalFormData: EggPopForm) => {
   const { data, error } = await browserClient.from("one_time_club").insert([finalFormData]).select("*").single();
   if (error) throw error;
   return data;
 };
 
 // supabase에 정기적 모임 제출
-export const submitRegularClubData = async (finalFormData: RegularClubForm) => {
+export const submitRegularClubData = async (finalFormData: EggClubForm) => {
   const { data, error } = await browserClient.from("regular_club").insert([finalFormData]).select("*").single();
   if (error) throw error;
   return data;
@@ -113,7 +108,7 @@ export const getAllRegularClubList = async () => {
 };
 
 // 모임장 정기적 모임 승인 테이블에 집어넣기
-export const putRepresentative = async (representative: RegularRequest) => {
+export const putRepresentative = async (representative: EggClubRequest) => {
   const { data, error } = await browserClient
     .from("r_c_participation_request")
     .insert([representative])
@@ -124,14 +119,14 @@ export const putRepresentative = async (representative: RegularRequest) => {
 };
 
 // 모임장 정기적 모임 맴버 테이블에 집어넣기
-export const putRegularMember = async (member: RegularMember) => {
+export const putRegularMember = async (member: EggClubMember) => {
   const { data, error } = await browserClient.from("r_c_member").insert([member]).select("*").single();
   if (error) throw error;
   return data;
 };
 
 // 모임장 일회성 모임 맴버 테이블에 집어넣기
-export const putOneTimeMember = async (member: OneTimeMember) => {
+export const putOneTimeMember = async (member: EggPopMember) => {
   const { data, error } = await browserClient.from("o_t_c_member").insert([member]).select("*").single();
   if (error) throw error;
   return data;
@@ -170,14 +165,14 @@ export const getRegularNotification = async (clubId: number) => {
 };
 
 // 정기모임의 공지 집어넣기
-export const submitRegularClubNotification = async (finalData: RegularClubNotification) => {
+export const submitRegularClubNotification = async (finalData: EggDay) => {
   const { data, error } = await browserClient.from("r_c_notification").insert([finalData]).select("*").single();
   if (error) throw error;
   return data;
 };
 
 // 정기적 공지 맴버로 집어넣기
-export const submitRegularMember = async (member: InsertNotificationMember) => {
+export const submitRegularMember = async (member: InsertEggDayMember) => {
   const { data, error } = await browserClient.from("r_c_notification_member").insert(member).select("*").single();
   if (error) throw error;
   return data;
@@ -226,8 +221,6 @@ export const getNotificationMember = async (notificationId: number | undefined) 
     .from("r_c_notification_member")
     .select(`*, user(user_name, user_profile_img)`)
     .eq("r_c_notification_id", notificationId);
-
-  // console.log("이태연!!!!!!", notificationId);
 
   if (error) throw error;
   return data;
