+++ src/app/(pages)/(club)/club/_components/oneTimeClub/AddressSearch.tsx
@@ -1,45 +1,10 @@
 "use client";
 
 import { useState, useEffect } from "react";
-import { OneTimeProps } from "../../_types/ClubForm";
+import { AddressData, DaumPostcodeData } from "@/types/address.types";
+import { EggPopProps } from "@/types/eggpop.types";
 
-declare global {
-  interface Window {
-    daum: {
-      Postcode: new (config: {
-        oncomplete: (data: {
-          userSelectedType: string;
-          roadAddress: string;
-          jibunAddress: string;
-          bname: string;
-          buildingName: string;
-          apartment: string;
-          zonecode: string;
-        }) => void;
-      }) => {
-        open: () => void;
-      };
-    };
-  }
-}
-
-interface AddressData {
-  zonecode: string; // 우편주소
-  address: string; // 기본주소
-  detailAddress: string; // 상세주소
-}
-
-interface DaumPostcodeData {
-  userSelectedType: string; // 사용자가 선택한 주소 타입
-  roadAddress: string; // 도로명 주소
-  jibunAddress: string; // 지번 주소
-  bname: string; // 법정동/법정리 이름
-  buildingName: string; // 건물명
-  apartment: string; // 아파트 여부
-  zonecode: string; // 우편번호
-}
-
-const AddressSearch = ({ formData, setFormData }: OneTimeProps) => {
+const AddressSearch = ({ formData, setFormData }: EggPopProps) => {
   const [addressData, setAddressData] = useState<AddressData>({
     zonecode: "",
     address: "",
