+++ src/app/(pages)/(club)/club/_components/oneTimeClub/ClubTitle.tsx
@@ -1,6 +1,6 @@
-import { OneTimeProps } from "../../_types/ClubForm";
+import { EggPopProps } from "@/types/eggpop.types";
 
-const ClubTitle = ({ formData, setFormData }: OneTimeProps) => {
+const ClubTitle = ({ formData, setFormData }: EggPopProps) => {
   return (
     <div>
       <h1 className="mb-4">모임의 제목을 작성해 주세요</h1>
