--- src/app/(pages)/(club)/club/regular-club-sub/[id]/create/[subId]/_types/notifictionInfo.ts
@@ -1,22 +1,0 @@
-export interface NotificaitonInfo {
-  r_c_notification_id: number;
-  user_id: string;
-  r_c_notification_name: string;
-  r_c_notification_content: string;
-  r_c_notification_date_time: string;
-  r_c_notification_location: string;
-  r_c_notification_tax: number;
-  r_c_notification_create_at: string;
-  r_c_id: number;
-  r_c_notification_image: string;
-}
-
-export interface NotificationMember {
-  r_c_notification_id: number;
-  r_c_member_id: number;
-  user_id: string;
-  user: {
-    user_name: string;
-    user_profile_img: string;
-  };
-}
