+++ src/app/(pages)/(club)/club/regular-club-sub/[id]/page.tsx
@@ -1,17 +1,11 @@
 import { fetchSubCategories, getRegularClubNotification, getRegularMember } from "../../_api/supabase";
-import { getRegularClub, Member, SCategory } from "./_types/Crews";
 import TabLayout from "./_components/TabLayout";
 import ClubHeader from "./_components/ClubHeader";
 import HomeContent from "./_components/HomeContent";
 import RegularNotification from "./_components/RegularNotification";
-import { InSertRegularClubNotification } from "./create/_types/subCreate";
-
-type CrewInfo = {
-  memberId: number;
-  userId: string;
-  userName: string;
-  userImage: string;
-};
+import { SubCategory } from "@/types/category.types";
+import { getEggClub, Member, MemberInfo } from "@/types/eggclub.types";
+import { InSertEggDay } from "@/types/eggday.types";
 
 const OneTimeClubSubpage = async ({ params }: { params: { id: string } }) => {
   const regularClubId = Number(params.id);
@@ -20,10 +14,10 @@ const OneTimeClubSubpage = async ({ params }: { params: { id: string } }) => {
     getRegularMember(regularClubId),
     getRegularClubNotification(regularClubId),
     fetchSubCategories()
-  ])) as [Member[], InSertRegularClubNotification[], SCategory[]];
+  ])) as [Member[], InSertEggDay[], SubCategory[]];
 
   // 클럽 정보만 추출
-  const clubInfo: getRegularClub = memberData[0]?.regular_club;
+  const clubInfo: getEggClub = memberData[0]?.regular_club;
   // console.log("클럽인포:", clubInfo);
 
   // 일치하는 카테고리 찾기
@@ -31,14 +25,14 @@ const OneTimeClubSubpage = async ({ params }: { params: { id: string } }) => {
   const stringCategory = matchCategory?.s_c_name;
 
   // 참여 크루 정보 추출
-  const crewMembers: CrewInfo[] = memberData.map((member) => ({
+  const crewMembers: MemberInfo[] = memberData.map((member) => ({
     memberId: member.r_c_member_id,
     userId: member.user_id,
     userName: member.user.user_name,
     userImage: member.user.user_profile_img
   }));
 
-  // console.log("Hmm...", crewMembers);
+  console.log("Hmm...", crewMembers);
   // 호스트 정보 추출
   const hostInfo = crewMembers.find((member) => member.userId === clubInfo.user_id);
