+++ src/types/category.types.ts
@@ -1,3 +1,9 @@
+// 대분류 카테고리 타입
+export interface MainCategory {
+  m_c_id: number;
+  m_c_name: string;
+}
+
 // 중분류 카테고리 타입
 export interface SubCategory {
   s_c_id: number;
